racing/requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-search-documents==11.6.0
duckduckgo-search==7.5.1
gradio==6.9.0
openai==1.61.0
pymupdf==1.25.3
python-dotenv==1.0.1